@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -31,9 +31,6 @@ dependencies:
   # Fontes usadas no sistema
   google_fonts: ^2.0.0
   
-  # Ícones 
-  flutter_icons: ^1.1.0
-
   # Dependências GETX
   get: ^4.0.0-nullsafety.2
   get_storage: ^2.0.1
@@ -50,7 +47,8 @@ dependencies:
   google_sign_in: ^5.0.0
 
   # Bottom Navigation Bar package
-  bubble_bottom_bar: ^1.2.0
+  bottom_navy_bar: ^6.0.0
+  #bubble_bottom_bar: ^1.2.0
 
   flutter_localizations:
     sdk: flutter
