@@ -2,10 +2,10 @@ import 'package:firebase_auth/firebase_auth.dart';
 
 class UserModel {
 
-  String id;
-  String name;
-  String email;
-  String urlimage;
+  String? id;
+  String? name;
+  String? email;
+  String? urlimage;
 
   UserModel({ this.id, this.name, this.email, this.urlimage });
 
