@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-class AppColors{
+class AppColors {
   static final Color themeColor = Color(0XFFE0E0E0); //Cor padrão geral do app
   static final Color white = Color(0XFFFFFFFF);
   static final Color black = Color(0XFF000000);
@@ -8,11 +8,8 @@ class AppColors{
   static final Color grey300 = Color(0XFFBDBDBD);
   static final Color grey800 = Color(0XFF424242);
   static final Color transparent = Colors.transparent;
+  static final Color investcolor = Color(0XFFF9A825);
+  static final Color backgroundColor = Color(0XFFEEEEEE);
+  static final Color incomeColor = Color(0XFF388E3C);
+  static final Color expenseColor = Color(0XFFE53935);
 }
-
-
-//const THEMECOLOR = Color(0XFFE0E0E0); //Cor padrão geral do app
-const BACKGROUNDCOLOR = Color(0XFFEEEEEE);
-const INCOMECOLOR = Color(0XFF388E3C);
-const EXPENSECOLOR = Color(0XFFE53935);
-const INVESTCOLOR = Color(0XFFF9A825);
