@@ -1,10 +1,3 @@
-import 'package:flutter/material.dart';
 import 'package:get/get.dart';
-import 'package:habito_invest_app/app/global/widgets/app_colors.dart';
-import 'package:intl/intl.dart';
 
-class CategoriesAddUpdateController extends GetxController {
-
-
-  
-}
+class CategoriesAddUpdateController extends GetxController {}
