@@ -2,31 +2,29 @@ import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:habito_invest_app/app/global/widgets/app_colors.dart';
 import 'package:habito_invest_app/app/global/widgets/constants.dart';
-import 'package:habito_invest_app/app/modules/categories/categoriesaddupdate/categoriesaddupdate_controller.dart';
 
-class CategoriesAddUpdatePage extends StatelessWidget{
-  final CategoriesAddUpdateController _controller = CategoriesAddUpdateController();
+class CategoriesAddUpdatePage extends StatelessWidget {
+  //final CategoriesAddUpdateController _controller = CategoriesAddUpdateController();
 
   @override
   Widget build(BuildContext context) {
-    
     return Scaffold(
       appBar: AppBar(
         backgroundColor: AppColors.themeColor,
         title: Text('Categorias'),
         actions: [
           IconButton(
-            onPressed: () { Get.back(); },
+            onPressed: () {
+              Get.back();
+            },
             icon: Icon(Icons.cancel, color: AppColors.black),
           ),
-
           IconButton(
-            onPressed: () { /*Código para salvar*/ },
+            onPressed: () {/*Código para salvar*/},
             icon: Icon(Icons.save, color: AppColors.black),
           ),
         ],
       ),
-
       body: Form(
         child: ListView(
           padding: EdgeInsets.symmetric(horizontal: 13.0),
@@ -35,25 +33,19 @@ class CategoriesAddUpdatePage extends StatelessWidget{
             TextFormField(
               style: TextStyle(fontWeight: FontWeight.bold),
               decoration: InputDecoration(
-                labelText: 'Nome',                
+                labelText: 'Nome',
               ),
             ),
             SizedBox(height: 5.0),
-
             TextFormField(
               style: TextStyle(fontWeight: FontWeight.bold),
               decoration: InputDecoration(
-                labelText: 'Observações',
-                alignLabelWithHint: true
-              ),
+                  labelText: 'Observações', alignLabelWithHint: true),
               keyboardType: TextInputType.multiline,
               maxLines: 5,
             ),
           ],
-
         ),
-        
-
       ),
     );
   }
