@@ -3,26 +3,22 @@ import 'package:get/get.dart';
 import 'package:habito_invest_app/app/global/widgets/app_colors.dart';
 import 'package:habito_invest_app/app/routes/app_routes.dart';
 
-class ExpenseList extends StatelessWidget{
-
+class ExpenseList extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      backgroundColor: BACKGROUNDCOLOR,
+      backgroundColor: AppColors.backgroundColor,
       appBar: AppBar(
         title: Text('Despesas'),
-        backgroundColor: EXPENSECOLOR,
+        backgroundColor: AppColors.expenseColor,
       ),
-      body: Center(
-        
-      ),
+      body: Center(),
       floatingActionButton: FloatingActionButton(
         onPressed: () => Get.toNamed(Routes.EXPENSE_ADDUPDATE),
-        backgroundColor: EXPENSECOLOR,
+        backgroundColor: AppColors.expenseColor,
         tooltip: 'Nova Despesa',
         child: Icon(Icons.add),
       ),
     );
   }
-
 }
