@@ -4,12 +4,10 @@ import 'package:habito_invest_app/app/global/widgets/app_text_styles.dart';
 import 'package:habito_invest_app/app/modules/goals/components/card_widget.dart';
 
 class GoalsPage extends StatelessWidget {
-  
   @override
   Widget build(BuildContext context) {
-
     return Scaffold(
-      backgroundColor: BACKGROUNDCOLOR,
+      backgroundColor: AppColors.backgroundColor,
       appBar: AppBar(
         brightness: Brightness.dark,
         title: Text('Definir Metas', style: AppTextStyles.appBarTextLight),
@@ -19,19 +17,30 @@ class GoalsPage extends StatelessWidget {
       body: Column(
         children: [
           Container(
-            padding: EdgeInsets.all(10.0),
-            child: Column(
-              children: [
-                CardWidget(goalName: 'Investimentos', goalValue: '30%', goalUniverse: 'do total das receitas'),
-                SizedBox(height: 5.0),
-                CardWidget(goalName: 'Gastos não essenciais', goalValue: 'R\$1.500,00', goalUniverse: 'por período'),
-                SizedBox(height: 5.0),
-                CardWidget(goalName: 'Gastos não essenciais', goalValue: 'R\$1.500,00', goalUniverse: 'por período'),
-                SizedBox(height: 5.0),
-                CardWidget(goalName: 'Gastos não essenciais', goalValue: 'R\$1.500,00', goalUniverse: 'por período'),
-              ],
-            )
-          ),
+              padding: EdgeInsets.all(10.0),
+              child: Column(
+                children: [
+                  CardWidget(
+                      goalName: 'Investimentos',
+                      goalValue: '30%',
+                      goalUniverse: 'do total das receitas'),
+                  SizedBox(height: 5.0),
+                  CardWidget(
+                      goalName: 'Gastos não essenciais',
+                      goalValue: 'R\$1.500,00',
+                      goalUniverse: 'por período'),
+                  SizedBox(height: 5.0),
+                  CardWidget(
+                      goalName: 'Gastos não essenciais',
+                      goalValue: 'R\$1.500,00',
+                      goalUniverse: 'por período'),
+                  SizedBox(height: 5.0),
+                  CardWidget(
+                      goalName: 'Gastos não essenciais',
+                      goalValue: 'R\$1.500,00',
+                      goalUniverse: 'por período'),
+                ],
+              )),
         ],
       ),
     );
