@@ -2,7 +2,7 @@ import 'package:get/get.dart';
 import 'package:habito_invest_app/app/data/model/user_model.dart';
 
 class HomeController extends GetxController {
-  final UserModel user = Get.arguments;
+  final UserModel? user = Get.arguments;
   RxInt _currentIndex = 0.obs;
   
   // Setter e Getter para variável currentIndex, para que a responsabilidade pela
