@@ -3,28 +3,22 @@ import 'package:get/get.dart';
 import 'package:habito_invest_app/app/global/widgets/app_colors.dart';
 import 'package:habito_invest_app/app/routes/app_routes.dart';
 
-class InvestList extends StatelessWidget{
-
+class InvestList extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
         title: Text('Investimentos'),
-        backgroundColor: INVESTCOLOR,
+        backgroundColor: AppColors.investcolor,
         iconTheme: IconThemeData(color: AppColors.black),
       ),
-      
-      body: Center(
-        
-      ),
-
+      body: Center(),
       floatingActionButton: FloatingActionButton(
         onPressed: () => Get.toNamed(Routes.INVEST_ADDUPDATE),
-        backgroundColor: INVESTCOLOR,
+        backgroundColor: AppColors.investcolor,
         tooltip: 'Novo Investimento',
         child: Icon(Icons.add, color: AppColors.black),
       ),
     );
   }
-
 }
